requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
djangorestframework-jwt<1.11.1,>=1.11.0
djangorestframework<3.11.1,>=3.11.0
django-rest-auth<0.9.6,>=0.9.5
django-filter
djangorestframework-gis